requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
joblib==1.3.2
numpy==1.26.3
packaging==23.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
scikit-learn==1.3.2
scipy==1.11.4
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.4
uritemplate==4.1.1
